Editing tests and adapters file
Moving tests

Adding required parameters in functions
