+++ drdown/users/adapters.py
@@ -7,10 +7,10 @@ from allauth.socialaccount.adapter import DefaultSocialAccountAdapter
 
 class AccountAdapter(DefaultAccountAdapter):
 
-    def is_open_for_signup(self):
+    def is_open_for_signup(self, request):
         return getattr(settings, 'ACCOUNT_ALLOW_REGISTRATION', True)
 
-    def get_email_confirmation_redirect_url(self):
+    def get_email_confirmation_redirect_url(self, request):
         return reverse('users:update')
 
     def get_login_redirect_url(self, request):
@@ -22,5 +22,5 @@ class AccountAdapter(DefaultAccountAdapter):
 
 class SocialAccountAdapter(DefaultSocialAccountAdapter):
 
-    def is_open_for_signup(self):
+    def is_open_for_signup(self, request):
         return getattr(settings, 'ACCOUNT_ALLOW_REGISTRATION', True)
