+++ drdown/users/tests/test_accounts.py
@@ -5,6 +5,11 @@ from django.urls import reverse
 from test_plus.test import TestCase
 from ..models import User
 from django.test.client import Client
+from allauth.account.models import (
+    EmailAddress,
+    EmailConfirmation,
+    EmailConfirmationHMAC,
+)
 
 
 class AccountsTestCase(TestCase):
@@ -16,12 +21,17 @@ class AccountsTestCase(TestCase):
         self.user = User.objects.create_user(username='test', password='12345')
         self.login_url = reverse('account_login')
         self.client = Client()
+        self.user2 = self.user = User.objects.create_user(
+            username='mariam', password='12345', name='Maria')
+        self.user3 = self.user = User.objects.create_user(
+            username='joao', password='12345')
 
     def test_login(self):
         """
         Test login
         """
-        response = self.client.post(self.login_url, {'username': 'test', 'password': '12345'})
+        response = self.client.post(
+            self.login_url, {'username': 'test', 'password': '12345'})
         self.assertEquals(response.status_code, 200)
 
     def login_successful(self):
@@ -32,7 +42,8 @@ class AccountsTestCase(TestCase):
         self.assertTrue(not response.wsgi_request.user.is_authenticated())
         data = {'username': 'test', 'password': '12345'}
         response = self.client.post(self.login_url, data)
-        redirect_url = reverse(AccountAdapter.get_login_redirect_url(self.request))
+        redirect_url = reverse(
+            AccountAdapter.get_login_redirect_url(self.request))
         self.assertEquals(response.status_code, 302)
         self.assertRedirects(response, redirect_url)
         self.assertTrue(response.wsgi_request.user.is_authenticated())
@@ -43,9 +54,23 @@ class AccountsTestCase(TestCase):
         """
         self.assertEqual(
             True,
-            AccountAdapter.is_open_for_signup(AccountAdapter),
+            AccountAdapter.is_open_for_signup(AccountAdapter, self.request),
         )
         self.assertEqual(
             True,
-            SocialAccountAdapter.is_open_for_signup(SocialAccountAdapter),
+            SocialAccountAdapter.is_open_for_signup(
+                SocialAccountAdapter, self.request),
         )
+
+    def test_redirect_login_sucess(self):
+        self.client.force_login(user=self.user2)
+        user_redirect = self.client.post('/accounts/login/')
+        detail_view = reverse('users:detail', kwargs={
+                              'username': self.user2.username})
+        self.assertRedirects(user_redirect, detail_view)
+
+    def test_redirect_not_logged_sucess(self):
+        self.client.force_login(user=self.user3)
+        user_redirect = self.client.post('/accounts/login/')
+        update_view = reverse('users:update')
+        self.assertRedirects(user_redirect, update_view)
